+++ src/views/ResourceList/DataListGroup/EmptyListGroupItem.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Col, ListGroupItem, Row } from 'reactstrap';
+
+const EmptyListGroupItem = () => (
+    <ListGroupItem tag="button" action onClick={() => {}}>
+        <Row>
+            <Col md={12}>
+                <h5>No matching records found.</h5>
+            </Col>
+        </Row>
+    </ListGroupItem>
+);
+
+export default EmptyListGroupItem;